// app/types/feature.d.ts
export type feature = {
    featureName: string,
    featureDescription: string,
    featureIcon: {
      id: string,
      alt: string,
      filename: string,
      mimetype: string,
      filesize: number,
      width: number,
      height: number,
      focalX: number,
      focalY: number,
      createdAt: string,
      updatedAt: string,
      url: string,
    }
  };